requirements.txt
base58==2.1.0
certifi==2021.5.30
charset-normalizer==2.0.3
dnspython==2.1.0
idna==3.2
more-itertools==8.8.0
py-bip39-bindings==0.1.7
py-ed25519-bindings==0.1.3
py-sr25519-bindings==0.1.3
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
scalecodec==0.11.16
six==1.16.0
substrate-interface==0.13.11
urllib3==1.26.6
websocket-client==0.59.0
xxhash==2.0.2
